# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST in Golang with Gin Framework",
        "contact": {
            "name": "Antonio Paya Gonzalez",
            "url": "http://antoniopg.tk",
            "email": "antonioalfa22@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/antonioalfa22/bms-monitoring/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/pipeline-evaluations": {
            "get": {
                "description": "Get pipeline evaluations by query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PipelineEvaluation"
                ],
                "summary": "Get pipeline evaluations by query",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Severity Critical Score",
                        "name": "severity_critical_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Severity High Score",
                        "name": "severity_high_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Severity Medium Score",
                        "name": "severity_medium_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Severity Low Score",
                        "name": "severity_low_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Threshold Score",
                        "name": "threshold_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_res.HTTPResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create pipeline evaluation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PipelineEvaluation"
                ],
                "summary": "Create pipeline evaluation",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.PipelineEvaluation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http_res.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/api/pipeline-evaluations/count": {
            "get": {
                "description": "Count pipeline evaluations by query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PipelineEvaluation"
                ],
                "summary": "Count pipeline evaluations by query",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Severity Critical Score",
                        "name": "severity_critical_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Severity High Score",
                        "name": "severity_high_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Severity Medium Score",
                        "name": "severity_medium_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Severity Low Score",
                        "name": "severity_low_score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Threshold Score",
                        "name": "threshold_score",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_res.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/api/pipeline-evaluations/{id}": {
            "get": {
                "description": "Get pipeline evaluation by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PipelineEvaluation"
                ],
                "summary": "Get pipeline evaluation by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_res.HTTPResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update pipeline evaluation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PipelineEvaluation"
                ],
                "summary": "Update pipeline evaluation by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.PipelineEvaluation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_res.HTTPResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete pipeline evaluation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PipelineEvaluation"
                ],
                "summary": "Delete pipeline evaluation by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_res.HTTPResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http_res.HTTPResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "data_count": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "users.PipelineEvaluation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "severity_critical_score": {
                    "type": "integer"
                },
                "severity_high_score": {
                    "type": "integer"
                },
                "severity_low_score": {
                    "type": "integer"
                },
                "severity_medium_score": {
                    "type": "integer"
                },
                "threshold_score": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}